Fix more encoding issues:
C++ does not have a standard encoding for its files, if we write
literals with accents, they are considered UTF-8 by some compilers and
they get values that are higher than what a `char` can handle. That's
why there was an unneeded cast in this block.

Using the Latin1 encoding the code has defined, we can deal with all
this in `char` range, without errors. I don't really understand why
everything is full of characters with tildes, if there's a fantastic
`chset.h` that contains all the magic needed to avoid these issues and
be portable...